test: wait for startup tab cleanup before asserting removal
